+++ config/initializers/bitsy.rb
@@ -13,6 +13,14 @@ Bitsy.configure do |c|
   # List of allowed secrets from Blockchain.info. The callback URL should look
   # like this http://app.com/bitsy/v1/blockchain_notifications?secret=mysecret
   c.blockchain_secrets = %w(secret)
+
+  # Write out the payment transactions, addresses, and other details. These
+  # transactions are the ones being forwarded.
+  c.send_many_log_path = Rails.root.join("log", "send_many.log")
+
+  # WHen debug is true, no money will be forwarded. This is for testing
+  # purposes.
+  c.debug = false
 end
 
 Sidekiq.configure_server do |config|